Extract class Gun to gun.ts as its growing
Add class QuadralBullet

In Gun extract creating bullet to method so that we can easily select which bullet pattern to be shoot
