+++ src/bullet.ts
@@ -35,3 +35,30 @@ export class DualBullet extends Phaser.Group {
         return this.children[0].y <= this.game.camera.y;
     }
 }
+
+export class QuadralBullet extends Phaser.Group {
+    constructor(game: Phaser.Game, x: number, y: number) {
+        super(game);
+
+        const bullets = createQuadralBullets(game, x, y);
+        this.add(bullets[0]);
+        this.add(bullets[1]);
+        this.add(bullets[2]);
+        this.add(bullets[3]);
+
+        game.add.existing(this);
+    }
+
+    update() {
+        this.setAll('body.velocity.y', -BULLET_VELOCITY);
+    }
+}
+
+export function createQuadralBullets(game: Phaser.Game, x: number, y: number): Bullet[] {
+    return [
+        new Bullet(game, x + 25, y + 30),
+        new Bullet(game, x + 50, y + 30),
+        new Bullet(game, x + (160 - 25), y + 30),
+        new Bullet(game, x + (160 - 50), y + 30)
+    ];
+}
